% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text_formatting.R
\name{add_info_icon}
\alias{add_info_icon}
\title{Add info-icon to text}
\usage{
add_info_icon(mytext, button_id)
}
\arguments{
\item{mytext}{the main text after which the info button will appear}

\item{button_id}{the id of info button that can be referenced by e.g. observeEvent}
}
\value{
html object
}
\description{
Add info-icon to text
}
\examples{
\dontrun{
mytext <- "Parameters"
infoButtonId <- "boxinfo"
add_info_icon(mytext, infoButtonId)
}
}
