% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_input_files.R
\name{check_forcings_ts}
\alias{check_forcings_ts}
\title{check if the imported forcings meet the requirements}
\usage{
check_forcings_ts(forcings, expected_forcings)
}
\arguments{
\item{forcings}{the imported forcings data in long format}

\item{expected_forcings}{the forcings required as a character vector}
}
\value{
NULL if requirements are met; error if not
}
\description{
check if the imported forcings meet the requirements
}
\examples{
\dontrun{
f_dat <- list(
               rad = data.frame(time = 1:10, value = rnorm(10)),
               temp = data.frame(time = 1:10, value = rnorm(10))
               )
exp_f <- c("rad", "temp")
check_forcings_ts(f_dat, exp_f)

exp_f <- c("irr", "tmp")
check_forcings_ts(f_dat, exp_f)
}
}
