% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model_helpers.R
\name{check_model_complete}
\alias{check_model_complete}
\title{Checks if the model parameters, forcings or initial values are complete}
\usage{
check_model_complete(x, type = "param")
}
\arguments{
\item{x}{the effect model object}

\item{type}{the type to check; either "param", "init" or forcings}
}
\description{
Checks if the model parameters, forcings or initial values are complete
}
\examples{
check_model_complete(cvasi::metsulfuron, type = "param")
check_model_complete(cvasi::metsulfuron, type = "init")
check_model_complete(cvasi::metsulfuron, type = "forcing")

check_model_complete(cvasi::GUTS_RED_SD(), type = "forcings")
}
