% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model_helpers.R
\name{construct_model}
\alias{construct_model}
\title{Call constructor of model}
\usage{
construct_model(model_name)
}
\arguments{
\item{model_name}{}
}
\description{
Call the constructor of a model by its name as "string"
}
\examples{
construct_model("GUTS_RED_IT")
}
