% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_custom_widgets.R
\name{create_valuebox}
\alias{create_valuebox}
\title{Function to create value boxes}
\usage{
create_valuebox(value, subtitle, width = 3)
}
\arguments{
\item{value}{to display in big bold letters}

\item{subtitle}{to display in smaller letter below 'value'}

\item{width}{width of the box}
}
\value{
a shinydashboard value box
}
\description{
Create based on the value (logical) a green thumbs-up box if `TRUE`, or a red 
thumbs-down box, if `FALSE`.
}
