% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model_helpers.R
\name{create_vb}
\alias{create_vb}
\title{Function to create value boxes}
\usage{
create_vb(value, subtitle, width = 3)
}
\arguments{
\item{value}{to display in big bold letters}

\item{subtitle}{to display in smaller letter below 'value'}

\item{width}{width of the box}
}
\description{
Function to create value boxes
}
