% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_custom_widgets.R
\name{dashboardbox_left}
\alias{dashboardbox_left}
\title{Create a dashboard box collapsible on click on title}
\usage{
dashboardbox_left(...)
}
\arguments{
\item{...}{arguments passed to [shinydashboard::box()]}
}
\value{
a shinydasboard box with collapse icon on the left side
}
\description{
Create a dashboard box collapsible on click on title
}
