% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_exposure}
\alias{default_exposure}
\title{Default exposure timeseries}
\format{
A data.frame with
\describe{
 \item{time}{time}
 \item{conc}{concentration}
 \item{trial}{indicates the trial of several trials with increasing application rates are used}
}
}
\usage{
default_exposure
}
\description{
Default exposure timeseries
}
\keyword{datasets}
