% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{empty_df}
\alias{empty_df}
\title{create an empty data.frame}
\usage{
empty_df(colnames)
}
\arguments{
\item{colnames}{the colnames of the returned data.frame}
}
\description{
create an empty data.frame
}
