% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text_formatting.R
\name{field_label}
\alias{field_label}
\title{Generate label for input fields with unit}
\usage{
field_label(name, unit)
}
\arguments{
\item{name}{the parameter/initial value name}

\item{unit}{the unit of the parameter/initial value}
}
\value{
a string of the parameter/initial value name and its unit
}
\description{
Generate label for input fields with unit
}
