% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model_helpers.R
\name{forcings_required}
\alias{forcings_required}
\title{Does the model require forcings?}
\usage{
forcings_required(x)
}
\arguments{
\item{x}{the effect model object}
}
\value{
TRUE if forcings are required, FALSE otherwise
}
\description{
Does the model require forcings?
}
\examples{
\dontrun{
forcings_required(cvasi::metsulfuron)
}
}
