% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_parameter_info.R
\name{get_parameter_info}
\alias{get_parameter_info}
\title{get info from parameter_descriptions data}
\usage{
get_parameter_info(model_, parameter_, type_)
}
\arguments{
\item{model}{}

\item{parameter}{}

\item{type}{one of "description", "unit", "group", "expert.value", "default"}
}
\description{
get info from parameter_descriptions data
}
\examples{
get_parameter_info(model_ = "Lemna_Schmitt", parameter_ = "Tmin", type_ ="description")
get_parameter_info(model_ = "Lemna_Schmitt", parameter_ = "Tmin", type_ ="unit")
get_parameter_info(model_ = "Lemna_Schmitt", parameter_ = "Tmin", type_ ="group")
get_parameter_info(model_ = "Lemna_Schmitt", parameter_ = "Tmin", type_ ="expert.value")
get_parameter_info(model_ = "Lemna_Schmitt", parameter_ = "Tmin", type_ ="default")
}
