% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_parameter_gouping.R
\name{group_title_with_icon}
\alias{group_title_with_icon}
\title{Create a title with icon for the thematic parameter groups}
\usage{
group_title_with_icon(group_title)
}
\arguments{
\item{group_title}{a string denoting the parameter group}
}
\value{
a shiny.tag object with specified icon or if group not defined the input string is returned unmodified
}
\description{
Create a title with icon for the thematic parameter groups
}
