% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_input_files.R
\name{import_forcings}
\alias{import_forcings}
\title{Import forcings from file}
\usage{
import_forcings(
  filepath,
  sep = ";",
  time_col = "time",
  forcing_col = "forcing",
  value_col = "value"
)
}
\arguments{
\item{filepath}{path to file with forcings data}

\item{sep}{value seperator}

\item{forcing_col}{columnname that contains the}
}
\value{
a list of forcings with a data.frame for each forcing
}
\description{
Import a text file with three c
}
