% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{in_range}
\alias{in_range}
\title{Validate if values are in range}
\usage{
in_range(value, min_value, max_value)
}
\arguments{
\item{value}{the values to validate}

\item{min_value}{the minimum value of the range}

\item{max_value}{the maximum value of the range}
}
\value{
TRUE if in range; FALSE otherwise
}
\description{
Function validates if values are in range. 'NA' are treated as no range bounds.
}
