% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{is_named_list}
\alias{is_named_list}
\title{Validate if list is named}
\usage{
is_named_list(lst)
}
\arguments{
\item{lst}{the list to validate}
}
\value{
TRUE if the list is named; FALSE otherwise
}
\description{
Validate if list is named
}
\examples{
\dontrun{
x1 <- list(a = 1:10, b = 2:5, c = 3)
is_named_list(x1) # TRUE

x2 <- list(1:10, 2:5, 3)
is_named_list(x2) # FALSE

x3 <- list(a= 1:10, 2:5, 3)
is_named_list(x3) # FALSE
}
}
