% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{list_rbind2}
\alias{list_rbind2}
\title{Extension of purrr::list_rbind}
\usage{
list_rbind2(
  x,
  listnames_to = "listname",
  ...,
  names_to = rlang::zap(),
  ptype = NULL
)
}
\arguments{
\item{x}{}

\item{listnames_to}{}

\item{...}{}

\item{names_to}{}

\item{ptype}{}
}
\description{
When list is bind the names of the list elements are added as additional column
}
\examples{
x1 <- list(a = data.frame(time=1:3, value = rnorm(3)),
           b = data.frame(time=1:3, value = rnorm(3)),
           c = data.frame(time=1:3, value = rnorm(3)))
list_rbind2(x1, listnames_to = "my_factor")
}
