% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model_helpers.R
\name{lookup_name}
\alias{lookup_name}
\title{Lookup model/scenario/contructor name}
\usage{
lookup_name(x, lookup_table = model_lookup, from = "scenario", to = "model_f")
}
\arguments{
\item{x}{the string to lookup}

\item{lookup_table}{the lookup table to use}

\item{from}{find x in column "from"}

\item{to}{get value in column "to}
}
\value{
a string with the found value in column "to"
}
\description{
Lookup model/scenario/contructor name
}
\examples{
\dontrun{
"LemnaSchmittScenario" \%>\% 
  lookup_name(model_lookup,
            from = "scenario",
            to = "model_f")
}
}
