% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_defaults}
\alias{model_defaults}
\title{Default values}
\format{
A list for each selectable model
\describe{
 \item{init_defaults}{Default values for model initial values}
 \item{parameter_defaults}{Default values for the model parameters}
 \item{forcing_defaults}{Default values for the model forcings}
}
}
\usage{
model_defaults
}
\description{
Default values
}
\keyword{datasets}
