% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_descriptions}
\alias{model_descriptions}
\title{Generic description of the models}
\format{
a list with an element for each model with a string describing the model
}
\usage{
model_descriptions
}
\description{
NB: This data set is planned to be removed and exchanged by an automated 
extraction of model descriptions directly from the CVASI documentation.#'
}
\keyword{datasets}
