% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_lookup}
\alias{model_lookup}
\title{Lookup table for model names, scenario/class names and constructor function names}
\format{
A data.frame with 
\describe{
  \item{model_name}{the model name that is returned from `get_name`}
  \item{scenario}{the scenario/class name that is returned from `class`}
  \item{model_f}{the name of the constructor function; listed in `model_choices`}
}
}
\usage{
model_lookup
}
\description{
Lookup table for model names, scenario/class names and constructor function names
}
\keyword{datasets}
