% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_parameter_gouping.R
\name{order_parameter_groups}
\alias{order_parameter_groups}
\title{Order parameter groups}
\usage{
order_parameter_groups(x, group_order)
}
\arguments{
\item{x}{a list with grouped parameter names; result from `group_parameters`}

\item{group_order}{a character vector indicating the order of parameter groups}
}
\value{
`x` ordered by `group_order`
}
\description{
Order parameter groups
}
\examples{
\dontrun{
parameter_groups <- list(
physiological = c("k_phot_fix", "k_phot_max", "k_resp", 
   "k_loss", "Tmin", "Tmax", "Topt", "t_ref", "Q10", "k_0", "a_k", 
   "C_P", "CP50", "a_P", "KiP", "C_N", "CN50", "a_N", "KiN", "BM50", 
   "mass_per_frond", "BMw2BMd"), 
toxicodynamic = c("Emax", "EC50", "b"), 
toxicokinetic = c("P_up", "AperBM", "Kbm", "P_Temp", "MolWeight"))

parameter_groups
order_parameter_groups(parameter_groups, c("toxicodynamic", "toxicokinetic", "physiological"))
}
}
