% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pack_unpack.R
\name{pack}
\alias{pack}
\title{Serialize a list to JSON/zip}
\usage{
pack(x, zipfile)
}
\arguments{
\item{x}{the list that is serialized}

\item{zipfile}{path to output zipfile}
}
\value{
path to the zip file
}
\description{
Serialize a list to JSON/zip
}
