% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameter_descriptions}
\alias{parameter_descriptions}
\title{Detailed information on the parameters and state variables of the available models}
\format{
a data.frame with
\describe{
 \item{model}{The 'name' of a model object}
 \item{parameter}{The parameter/state variable name}
 \item{description}{A description of the parameter/state variable}
 \item{unit}{The unit of the parameter/state variable}
 \item{group}{Grouping factor for the parameters 
             (with levels 'physiological', 'toxicodynamic' and 'toxicokinetic') and
             the state variable ('state variable)}
 \item{expert.value}{Is this a parameter/state variable that is suggested to be 
             modified only by experts? 'yes' or 'no'}
 \item{default}{A default value}
}
}
\usage{
parameter_descriptions
}
\description{
Detailed information on the parameters and state variables of the available models
}
\keyword{datasets}
