% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_exposure_profiles.R
\name{read_exposure_profiles}
\alias{read_exposure_profiles}
\title{Read exposure profiles}
\usage{
read_exposure_profiles(pathtofiles, profileNames = NA)
}
\arguments{
\item{pathtofiles}{}

\item{profileNames}{}
}
\value{
list of exposure profiles
}
\description{
Read several exposure profiles at their corresponding pathes at 'pathtofiles'
}
