% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_exposure_profiles.R
\name{read_single_exposure}
\alias{read_single_exposure}
\title{Read an exposure profile}
\usage{
read_single_exposure(pathtofile, sep = "\\t")
}
\arguments{
\item{pathtofile}{the path where the exposure profile is found}
}
\value{
the exposure profile
}
\description{
Read a single exposure profile at 'pathtofile'
}
