% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_custom_widgets.R
\name{set_lang}
\alias{set_lang}
\title{Set locale of input container}
\usage{
set_lang(x, locale = "en-US")
}
\arguments{
\item{x}{result of `shiny::numericInput()`}

\item{locale}{locale as string, e.g. "de-DE" for German locale}
}
\description{
Set locale of input container
}
