% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_templates.R
\name{template_exposure}
\alias{template_exposure}
\title{create exposure file template}
\usage{
template_exposure(filepath, trials = FALSE, example_data = FALSE)
}
\arguments{
\item{filepath}{file destination; use txt file ending.}

\item{trials}{should a 'trial' column be created?}

\item{example_data}{should the data contain example values? 
If FALSE only NA will be included}
}
\value{
the path to the example file if a filepath was given. 
If filepath is missing the data is returned as data.frame
}
\description{
create exposure file template
}
