% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_templates.R
\name{template_forcings}
\alias{template_forcings}
\title{create exposure file template}
\usage{
template_forcings(filepath, forcings)
}
\arguments{
\item{filepath}{file destination; use txt file ending.}

\item{forcings}{a vector of strings indicating the}
}
\value{
the path to the example file if a filepath was given. 
If filepath is missing the data is returned as data.frame
}
\description{
create exposure file template
}
