% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_description.R
\name{toHTML}
\alias{toHTML}
\title{Convert to md to html}
\usage{
toHTML(x)
}
\arguments{
\item{x}{a string or a vector of strings}
}
\value{
html character
}
\description{
Convert to md to html
}
\examples{
\dontrun{
s <- c("This ", "is not ", "a test")
toHTML(s)
}
}
