% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text_formatting.R
\name{tooltip_text}
\alias{tooltip_text}
\title{create an div container with title}
\usage{
tooltip_text(mytext, tooltip)
}
\arguments{
\item{mytext}{the text to be displayed}

\item{tooltip}{The text that appears when the mouse hovers over `mytext`}
}
\value{
an html div-container
}
\description{
create an div container with title
}
\examples{
\dontrun{
tooltip_text(mytext = "Tmin", 
             tooltip = get_parameter_info(
                          model_ = "Lemna_Schmitt", 
                          parameter_ = "Tmin", 
                          type_ ="description"
                          )
             )
tooltip_text(mytext = "Tmin")
}
}
