% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pack_unpack.R
\name{unpack}
\alias{unpack}
\title{Deserialize a list to JSON/zip}
\usage{
unpack(zipfile)
}
\arguments{
\item{zipfile}{path to output zipfile}
}
\value{
the content of the zipped json file as list
}
\description{
Deserialize a list to JSON/zip
}
