% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model_helpers.R
\name{check_exposure_complete}
\alias{check_exposure_complete}
\title{Checks if the exposure data is complete}
\usage{
check_exposure_complete(x)
}
\arguments{
\item{x}{}
}
\description{
Checks if the exposure data is complete
}
\examples{
exp_dat <- neofm::Schmitt2013 |>
               dplyr::select(t, ID, conc) |>
               dplyr::rename(trial = "ID", time = "t")
check_exposure_complete(exp_dat)
}
