% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model_helpers.R
\name{check_forcings_complete}
\alias{check_forcings_complete}
\title{Checks if the forcing data is complete}
\usage{
check_forcings_complete(expected_forcings, forcings)
}
\arguments{
\item{expected_forcings}{a character vector listing the expected forcings}

\item{forcings}{a character vector listing the available forcings}
}
\value{
TRUE if forcings are complete, FALSE otherwise
}
\description{
Checks if the forcing data is complete
}
