% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_model_helpers.R
\name{get_required}
\alias{get_required}
\title{Get the required parameter of init names of a model}
\usage{
get_required(x, type)
}
\arguments{
\item{x}{the effect model object}

\item{type}{the type to check; either "param", "init" or forcing}
}
\description{
Get the required parameter of init names of a model
}
\examples{
dat <- neofm::GUTS_RED_IT()
get_required(dat, type = "param")
get_required(dat, type = "init")
get_required(dat, type = "forcings")
}
