% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_description.R
\name{read_roxygen}
\alias{read_roxygen}
\title{Read roxygen help}
\usage{
read_roxygen(package, f_name, tag, print_tags_only = FALSE)
}
\arguments{
\item{package}{the name of the package}

\item{f_name}{the name of the function you want to find}

\item{tag}{the tag of the paragraph you want to extract}

\item{print_tags_only}{if TRUE the available tags are printed}
}
\value{
a string of the content of the respective roxygen header tag or a vector with available tags of a function
}
\description{
Read roxygen help
}
\examples{
\dontrun{
read_roxygen(package = "base", f_name = "data.frame", tag = "\\\\details")
}
}
